use winsock's struct timeval in the msvc sys/time.h shim

+++ crypto/compat/posix_win.c
@@ -303,7 +303,7 @@ int gettimeofday(struct timeval *tp, void *tzp)
 	time = ((uint64_t)file_time.dwLowDateTime);
 	time += ((uint64_t)file_time.dwHighDateTime) << 32;
 
-	tp->tv_sec = (long long)((time - EPOCH) / 10000000L);
+	tp->tv_sec = (long)((time - EPOCH) / 10000000L);
 	tp->tv_usec = (long)(system_time.wMilliseconds * 1000);
 	return 0;
 }
